enhance: expand e2e tests for koulutussopimus approval process with verification tokens
