+++ e2e/cypress/plugins/db-tasks/kouluttaja.ts
@@ -1,88 +1,15 @@
-import { randomUUID } from 'crypto'
 import type { Client } from 'pg'
 import { dbClient, withDb } from './db-client'
+import {
+  fetchUserIdsByEmailOrLogin,
+  ensureUserAuthority,
+  createVerificationToken,
+  createUserWithRole,
+  cleanupUser,
+} from './user-utils'
+import { ensureYliopistoErikoisalaLink } from './kayttaja-utils'
 
-async function createVerificationToken(client: Client, userId: string): Promise<string> {
-  const token = randomUUID()
-  await client.query(`DELETE FROM verification_token WHERE user_id = $1`, [userId])
-  await client.query(
-    `INSERT INTO verification_token (id, user_id) VALUES ($1, $2)`,
-    [token, userId]
-  )
-  return token
-}
-
-// ─── Helpers ──────────────────────────────────────────────────────────────────
-
-async function fetchKouluttajaIds(
-  client: Client,
-  email: string
-): Promise<{ userId: string; kayttajaId: number } | null> {
-  // Search by email OR login: the app may nullify the email column when the
-  // kouluttaja approves a form with an empty sahkoposti field in the DTO,
-  // so login (immutable, set to email at creation) is a reliable fallback key.
-  const result = await client.query(
-    `SELECT u.id AS user_id, k.id AS kayttaja_id
-     FROM jhi_user u
-     LEFT JOIN kayttaja k ON k.user_id = u.id
-     WHERE u.email = $1 OR u.login = $1`,
-    [email]
-  )
-  if (result.rows.length === 0) return null
-  return { userId: result.rows[0].user_id, kayttajaId: result.rows[0].kayttaja_id }
-}
-
-async function ensureYliopistoErikoisalaLink(client: Client, kayttajaId: number): Promise<void> {
-  await client.query(
-    `INSERT INTO kayttaja_yliopisto_erikoisala (kayttaja_id, yliopisto_id, erikoisala_id)
-     SELECT $1, 1, 46
-     WHERE NOT EXISTS (
-       SELECT 1 FROM kayttaja_yliopisto_erikoisala
-       WHERE kayttaja_id = $1 AND yliopisto_id = 1 AND erikoisala_id = 46
-     )`,
-    [kayttajaId]
-  )
-}
-
-async function ensureUserAuthority(client: Client, userId: string): Promise<void> {
-  await client.query(
-    `INSERT INTO jhi_user_authority (user_id, authority_name)
-     VALUES ($1, 'ROLE_KOULUTTAJA') ON CONFLICT DO NOTHING`,
-    [userId]
-  )
-}
-
-async function createKouluttajaUser(
-  client: Client,
-  email: string,
-  etunimi: string,
-  sukunimi: string
-): Promise<number> {
-  const userId = randomUUID()
-  const nimi = `${etunimi} ${sukunimi}`
-
-  await client.query(
-    `INSERT INTO jhi_user
-       (id, login, first_name, last_name, email, activated, lang_key,
-        created_by, last_modified_by, active_authority)
-     VALUES ($1,$2,$3,$4,$5,true,'fi','system','system','ROLE_KOULUTTAJA')`,
-    [userId, email, etunimi, sukunimi, email]
-  )
-  await ensureUserAuthority(client, userId)
-
-  const kayttajaId: number = (
-    await client.query(
-      `INSERT INTO kayttaja (nimike, user_id, tila)
-       VALUES ($1,$2,'AKTIIVINEN')
-       RETURNING id`,
-      [nimi, userId]
-    )
-  ).rows[0].id
-
-  await ensureYliopistoErikoisalaLink(client, kayttajaId)
-
-  return kayttajaId
-}
+const ROLE = 'ROLE_KOULUTTAJA'
 
 // ─── Exported tasks ───────────────────────────────────────────────────────────
 
@@ -96,70 +23,68 @@ async function createKouluttajaUser(
  */
 export const kouluttajaTasks = {
   async 'db:seedKouluttaja'({
-                              email,
-                              etunimi,
-                              sukunimi,
-                            }: {
+    email,
+    etunimi,
+    sukunimi,
+  }: {
     email: string
     etunimi: string
     sukunimi: string
   }): Promise<{ kayttajaId: number; token?: string } | null> {
-    return withDb(dbClient, async (client: any) => {
-      // JPA's @ManyToMany joins on jhi_authority.name, so this row must exist
-      // or JPQL queries won't find the kouluttaja (the left join produces NULL).
+    return withDb(dbClient, async (client: Client) => {
       await client.query(
-        `INSERT INTO jhi_authority (name) VALUES ('ROLE_KOULUTTAJA') ON CONFLICT DO NOTHING`
+        `INSERT INTO jhi_authority (name) VALUES ($1) ON CONFLICT DO NOTHING`,
+        [ROLE]
       )
 
-      const existing = await fetchKouluttajaIds(client, email)
-
+      const existing = await fetchUserIdsByEmailOrLogin(client, email)
       if (existing) {
-        // User already exists — ensure both the yliopisto link and authority row
-        // are present in case a previous partial run left them missing.
         await ensureYliopistoErikoisalaLink(client, existing.kayttajaId)
-        await ensureUserAuthority(client, existing.userId)
+        await ensureUserAuthority(client, existing.userId, ROLE)
         return { kayttajaId: existing.kayttajaId }
       }
 
-      const kayttajaId = await createKouluttajaUser(client, email, etunimi, sukunimi)
-      const created = await fetchKouluttajaIds(client, email)
+      const kayttajaId = await createUserWithRole(client, {
+        email,
+        etunimi,
+        sukunimi,
+        role: ROLE,
+        linkFn: ensureYliopistoErikoisalaLink,
+      })
+      const created = await fetchUserIdsByEmailOrLogin(client, email)
       const token = await createVerificationToken(client, created!.userId)
       return { kayttajaId, token }
     })
   },
 
-  /**
-   * Poistaa esialustetun kouluttajan tiedot. Turvallinen kutsua, vaikka käyttäjää
-   * ei olisikaan.
-   */
   async 'db:cleanupKouluttaja'({ email }: { email: string }): Promise<null> {
-    return withDb(dbClient, async (client: any) => {
-      const ids = await fetchKouluttajaIds(client, email)
+    return withDb(dbClient, async (client: Client) => {
+      const ids = await fetchUserIdsByEmailOrLogin(client, email)
       if (!ids) return null
       const { userId, kayttajaId } = ids
-
-      if (kayttajaId) {
-        await client.query(
-          `DELETE FROM rel_kayttaja_yliopisto_erikoisala__tehtavatyyppi
-           WHERE kayttaja_yliopisto_erikoisala_id IN (
-             SELECT id FROM kayttaja_yliopisto_erikoisala WHERE kayttaja_id = $1
-           )`,
-          [kayttajaId]
-        )
-        await client.query(
-          `DELETE FROM kayttaja_yliopisto_erikoisala WHERE kayttaja_id = $1`,
-          [kayttajaId]
-        )
-        await client.query(
-          `DELETE FROM kouluttajavaltuutus WHERE valtuutettu_id = $1`,
-          [kayttajaId]
-        )
-        await client.query(`DELETE FROM kayttaja WHERE id = $1`, [kayttajaId])
-      }
-
-      await client.query(`DELETE FROM jhi_user_authority WHERE user_id = $1`, [userId])
-      await client.query(`DELETE FROM verification_token WHERE user_id = $1`, [userId])
-      await client.query(`DELETE FROM jhi_user WHERE id = $1`, [userId])
+      // Clean up tehtavatyyppi, kayttaja_yliopisto_erikoisala, kouluttajavaltuutus, kayttaja
+      await cleanupUser(
+        client,
+        userId,
+        kayttajaId,
+        async (client, kayttajaId) => {
+          await client.query(
+            `DELETE FROM rel_kayttaja_yliopisto_erikoisala__tehtavatyyppi
+             WHERE kayttaja_yliopisto_erikoisala_id IN (
+               SELECT id FROM kayttaja_yliopisto_erikoisala WHERE kayttaja_id = $1
+             )`,
+            [kayttajaId]
+          )
+          await client.query(
+            `DELETE FROM kayttaja_yliopisto_erikoisala WHERE kayttaja_id = $1`,
+            [kayttajaId]
+          )
+          await client.query(
+            `DELETE FROM kouluttajavaltuutus WHERE valtuutettu_id = $1`,
+            [kayttajaId]
+          )
+        }
+      )
       return null
     })
   },
