+++ e2e/package.json
@@ -12,7 +12,7 @@
   },
   "devDependencies": {
     "@types/pg": "^8.11.0",
-    "cypress": "^15.13.0",
+    "cypress": "^15.14.0",
     "pg": "^8.13.0",
     "typescript": "^6.0.0"
   }
